separate ethereum and solana tests
